// src/centerpanel/Flows/types/LorazepamFormState.ts
export interface LorazepamFormState {

  indicationContext: string;



  preDoseMentalStatus: string;

  preDoseMotorFindings: string;

  preDoseVitalsAirway: string;



  lorazepamDetails: string;

  immediateResponse: string;

  safetyObservations: string;



  followupMonitoringPlan: string;

  reassessmentNeeds: string;

}